provide `iso_oscar_gap` and `iso_gap_oscar` for number fields
- provide `iso_oscar_gap` and `iso_gap_oscar` for number fields
  that are not of cyclotomic type
  (addresses the problem from #1930,
  but does not address the questions raised in its discussion)

- provide docstrings for `Oscar.iso_oscar_gap` and `Oscar.iso_gap_oscar`
  (had been suggested by Meinolf; it is not clear to me where in the
  manual these docstrins should appear)

- removed a workaround for certain finite fields, since the problem
  on the GAP side has been fixed
  (a test and an assertion make sure that the problem will not come up again)

- extended the tests

- replaced `GAP.Globals` access by `Oscar.GAPWrap` access

+++ src/GAP/gap_to_oscar.jl
@@ -44,7 +44,7 @@ function (F::FinField)(x::GAP.FFE)
         # FFE in GAP only exist for "small" characteristic, so we know the int
         # value fits into an Int; telling Julia about this via a type assertion
         # results in slightly better code
-        val = GAPWrap.INT_FFE_DEFAULT(x)::Int
+        val = GAPWrap.INT_FFE_DEFAULT(x)
         return F(val)
     end
